Change GraphQL types and mutations
Add a type check for pages and steps so it can only have workflow,
dashboards or forms in it.
Add an applications attribute in usertype so we can easily list all the
applications available for a specific user.
Correct some typos and add a new mutation to delete pages dynamically.