fix: When an application is duplicated, the access settings are not being copied.
The problem occurred because the new application inherited the role ids from the old application. This didn't work because the role is linked directly to the application. To resolve this, I changed the page duplication sequence, moved the creation of the roles of my new application to a later stage, and created a hash that links the roles of the old and new application. When duplicating the page, I check which page it corresponds to and create it correctly.