Refactor proxyAPIRequest and apiConfiguration.query.ts
- Refactored proxyAPIRequest to include skipCache parameter
- Added console.log statement in proxyAPIRequest
- Added conditional statement in apiConfiguration.query.ts for skipCache parameter

feat: Add TODO for handling skipCache argument in proxy API request

This commit adds a TODO comment in the code to handle the skipCache argument in the proxy API request. This will allow us to fetch data from the API even if skipCache is set to true.

Add support for skipping the cache in proxyAPIRequest function

This commit adds a new parameter `skipCache` to the `proxyAPIRequest` function. This allows us to specify whether we want to skip the cache when making an API request.

.